app/requirements.txt
asn1crypto
certifi
cffi
chardet
Click
cryptography
Flask
future
idna
itsdangerous
Jinja2
MarkupSafe
pycparser
python-telegram-bot
requests
six
urllib3
Werkzeug
